recipe_platform/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy>=2.0.36
pydantic[email]>=2.10.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
pydantic-settings>=2.6.0
bleach>=6.1.0
reportlab>=4.2.0
requests>=2.32.0
pytest>=8.3.0
pytest-asyncio>=0.24.0
httpx>=0.28.0
python-dotenv>=1.0.0
